feat: separate the config and initialize the context
