+++ .env.test
@@ -2,3 +2,4 @@ DATABASE_URL: mysql://root:root@field-seating-test-db:3306/field_seating_test
 JWT_SECRET=jwt_secret
 SIB_KEY="sibKey"
 EMAIL=YOUR_EMAIL
+REDIS_URL=
