+++ env.example
@@ -3,3 +3,4 @@ EMAIL=YOUR_EMAIL
 DATABASE_URL="mysql://[yourname]:[yourpassword]@localhost:[yourport]/field_seating_development"
 JWT_SECRET=
 SIB_KEY=
+REDIS_URL=redis://john:password@awesome.redis.server:6379
