--- src/config/config.json
@@ -1,23 +1,0 @@
-{
-  "development": {
-    "username": "root",
-    "password": "password",
-    "database": "field_seating_development",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  },
-  "test": {
-    "username": "root",
-    "password": "password",
-    "database": "field_seating_test",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  },
-  "production": {
-    "username": "root",
-    "password": "password",
-    "database": "field_seating_production",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  }
-}
