+++ src/routes/index.js
@@ -8,6 +8,8 @@ const user = require('./modules/user');
 const GeneralError = require('../errors/error/general-error');
 const signUpErrorMap = require('../errors/sign-up-error');
 const alwaysThrow = require('../utils/func/always-throw');
+const { isDevelopment } = require('../context');
+
 const router = express.Router();
 
 // use by signIn
@@ -33,7 +35,7 @@ router.use('/api/users', user);
 router.patch('/api/verify-email', userController.verifyEmail);
 
 // 檢視email格式使用
-if (process.env.NODE_ENV !== 'production') {
+if (isDevelopment()) {
   router.use('/testemail', (req, res) => res.render('verify-email'));
 }
